requirements.txt
pylint
pytest
yanc
nose
Pillow
